requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
contourpy==1.3.0
cycler==0.12.1
Django==4.2.16
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
fonttools==4.54.1
future==1.0.0
gunicorn==23.0.0
inflection==0.5.1
kiwisolver==1.4.7
Markdown==3.7
matplotlib==3.9.2
numpy==2.0.2
packaging==24.1
pillow==10.4.0
PyJWT==2.10.1
pyparsing==3.1.4
pyswarms==1.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scipy==1.13.1
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
uritemplate==4.1.1
wheel==0.44.0  # Fixed to use the PyPI version
